backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.3.0
gunicorn==22.0.0
whitenoise==6.6.0
